Trying to simplify some things.
SO currently it will seem to accept ONE connection and recieve
stuff from it and then die with an error.  Not sure what the heck
is up.  But the only difference I can see between it and the
quinn server example code is quinn spawns a new future per stream
and this tries to handle them all in one go, so let's change
that and see what happens...