Sloooowly merging quinn's networking in.
Works so far, think this can work.

Again, the difference seems to be it just spawns lots of
futures to Do Stuff rather than trying to chain futures together,
which is probably better.

We use current_thread::spawn() to spawn things instead
of the peer's `Runtime`, which seems fine; we can't clone
or send the `Runtime` between futures, though we can create
`Handles` from it which we can do that for.  But if we can
just use `current_thread::spawn()` and it works then we can
just heckin' do that.